pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentcli"
version = "1.0.0"
description = "A minimal CLI coding agent with streaming + tool calls"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Chirag Jain"}]
dependencies = [
  "litellm>=1.44.0",
  "rich>=13.7.1",
  "python-dotenv>=1.0.1",
  "typer>=0.12.3",
  "ddgs>=7.0.0",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.3",
]

[project.scripts]
agentcli = "agentcli.cli:app"

[tool.setuptools]
packages = ["agentcli", "agentcli.tools"]
